:art: code architectuire and model are working !!!
